change gImageCount to uint32_t

+++ sphere_forge/main.cpp
@@ -29,7 +29,7 @@ struct UniformBlock {
 };
 
 Renderer *pRenderer = nullptr;
-constexpr size_t gImageCount = 3;
+constexpr uint32_t gImageCount = 3;
 
 constexpr int gSphereResolution = 10;
 constexpr float gSphereDiameter = 1.0f;
